+++ packages/plugins/Draw/src/types.ts
@@ -22,7 +22,7 @@ export type Mode = 'none' | 'draw' | 'edit' | 'delete'
 
 export interface CreateInteractionsPayload {
   drawSource: VectorSource
-  drawLayer: VectorLayer<Feature>
+  drawLayer: VectorLayer
 }
 
 export interface DrawState {