+++ packages/plugins/AddressSearch/src/index.ts
@@ -2,7 +2,7 @@ import Vue from 'vue'
 import { AddressSearchConfiguration } from '@polar/lib-custom-types'
 
 import { AddressSearch } from './components'
-import language from './language'
+import locales from './locales'
 import { makeStoreModule } from './store'
 
 import SearchResultSymbols from './utils/searchResultSymbols'
@@ -18,7 +18,7 @@ export default (options: AddressSearchConfiguration) => (instance: Vue) =>
   instance.$store.dispatch('addComponent', {
     name: 'addressSearch',
     plugin: AddressSearch,
-    language,
+    locales,
     storeModule: makeStoreModule(),
     options,
   })