+++ packages/plugins/LoadingIndicator/src/index.ts
@@ -1,14 +1,14 @@
 import Vue from 'vue'
 import { PluginOptions } from '@polar/lib-custom-types'
 import { LoadingIndicator } from './components'
-import language from './language'
+import locales from './locales'
 import { makeStoreModule } from './store'
 
 export default (options: PluginOptions) => (instance: Vue) =>
   instance.$store.dispatch('addComponent', {
     name: 'loadingIndicator',
     plugin: LoadingIndicator,
-    language,
+    locales,
     storeModule: makeStoreModule(),
     options,
   })