+++ packages/plugins/Pins/src/index.ts
@@ -1,12 +1,12 @@
 import Vue from 'vue'
 import { PinsConfiguration } from '@polar/lib-custom-types'
 import { makeStoreModule } from './store'
-import language from './language'
+import locales from './locales'
 
 export default (options: PinsConfiguration) => (instance: Vue) =>
   instance.$store.dispatch('addComponent', {
     name: 'pins',
-    language,
+    locales,
     storeModule: makeStoreModule(),
     options,
   })