+++ packages/plugins/Toast/src/index.ts
@@ -1,14 +1,14 @@
 import Vue from 'vue'
 import { ToastConfiguration } from '@polar/lib-custom-types'
 import { Toast } from './components'
-import language from './language'
+import locales from './locales'
 import { makeStoreModule } from './store'
 
 export default (options: ToastConfiguration) => (instance: Vue) =>
   instance.$store.dispatch('addComponent', {
     name: 'toast',
     plugin: Toast,
-    language,
+    locales,
     options,
     storeModule: makeStoreModule(),
   })