+++ packages/types/custom/core.ts
@@ -520,14 +520,16 @@ export interface ZoomConfiguration extends PluginOptions {
   showZoomSlider?: boolean
 }
 
-export interface LanguageOption {
+export interface Locale {
   resources: Resource
   /** Language key as described in the i18next documentation */
   type: string
 }
 
 export interface PluginContainer {
-  language: LanguageOption[]
+  locales: Locale[]
+  /** @deprecated Please use .locales instead */
+  language: Locale[]
   name: string
   options: PluginOptions
   plugin: object
@@ -669,7 +671,7 @@ export interface MapConfig extends MasterportalApiConfig {
   extendedMasterportalapiMarkers?: ExtendedMasterportalapiMarkers
   featureStyles?: string
   language?: InitialLanguage
-  locales?: LanguageOption[]
+  locales?: Locale[]
   renderFaToLightDom?: boolean
   stylePath?: string
   vuetify?: UserVuetifyPreset